[Ide] Split projection data into own class
